clear essensplan before removeEssensplan

+++ src/main/java/de/leuphana/mensaplus/service/EssensplanService.java
@@ -40,7 +40,12 @@ public class EssensplanService {
     }
 
     public void deleteEssensplanById(int id) {
-        essensplanRepository.deleteById(id);
+        Optional<Essensplan> optionalEssensplan = essensplanRepository.findById(id);
+        if (optionalEssensplan.isPresent()) {
+            Essensplan essensplan = optionalEssensplan.get();
+            essensplan.getEssenProWoche().clear();
+            essensplanRepository.deleteById(id);
+        }
     }
 
     @Transactional
